Fix assorted medium-severity bugs across multiple components
- template text: clamp flash-stored length to buffer size to prevent
  OOB read on corrupted preference data
- nfc: validate language_code_length against payload size before
  constructing strings from NDEF text records
- nfc: bounds-check TLV index from get_mifare_classic_ndef_start_index
  before accessing data vector (return values 254/255 caused OOB)
- ina2xx_base: fix off-by-one in two_complement_ and simplify to
  shift-based sign extension
- msa3xx: same fix and simplification in twos_complement_
- bytebuffer: put_uint24 with offset wrote 4 bytes (sizeof uint32_t)
  instead of 3; use put_uint32_ with explicit length like put_int24
- cap1188: write_register passed length 2 for 1-byte variable,
  reading stack garbage into adjacent register
- hte501: CRC validation used && instead of ||, silently accepting
  data when only one of two CRC checks failed
- inkplate: uint8_t clock truncated for GPIO pins >= 8; use uint32_t
  to match GPIO.out_w1ts type

+++ esphome/components/inkplate/inkplate.cpp
@@ -704,7 +704,7 @@ void Inkplate::vscan_start_() {
 }
 
 void Inkplate::hscan_start_(uint32_t d) {
-  uint8_t clock = (1 << this->cl_pin_->get_pin());
+  uint32_t clock = (1 << this->cl_pin_->get_pin());
   this->sph_pin_->digital_write(false);
   GPIO.out_w1ts = d | clock;
   GPIO.out_w1tc = this->get_data_pin_mask_() | clock;
